Move SearchBox and related components from navigation to search directory
