+++ packages/components/releaseNotes/components.md
@@ -3,9 +3,10 @@ Components, models, actions, and utility functions for LabKey applications and p
 
 ### version TBD
 *Released*: TBD
-* Foundational work for Sample Finder feature devleopment.
+* Foundational work for Sample Finder feature development.
   * Add experimental feature flag
   * Update FindAndSearch menu
+  * Move SearchBox and related componets into the search directory
 
 ### version 2.111.0
 *Released*: 24 December 2021
