+++ packages/components/src/internal/components/navigation/NavigationBar.spec.tsx
@@ -31,7 +31,7 @@ import { MenuSectionModel, ProductMenuModel } from './model';
 import { NavigationBar } from './NavigationBar';
 import { UserMenu } from './UserMenu';
 import { ProductMenu } from './ProductMenu';
-import { SearchBox } from './SearchBox';
+import { SearchBox } from '../search/SearchBox';
 
 beforeEach(() => {
     LABKEY.devMode = false;
