+++ packages/components/src/internal/components/navigation/NavigationBar.tsx
@@ -26,11 +26,11 @@ import { ProductNavigation } from '../productnavigation/ProductNavigation';
 import { shouldShowProductNavigation } from '../productnavigation/utils';
 
 import { ProductMenu } from './ProductMenu';
-import { SearchBox } from './SearchBox';
+import { SearchBox } from '../search/SearchBox';
 import { UserMenu, UserMenuProps } from './UserMenu';
 import { MenuSectionConfig } from './ProductMenuSection';
 import { ProductMenuModel } from './model';
-import { FindAndSearchDropdown } from './FindAndSearchDropdown';
+import { FindAndSearchDropdown } from '../search/FindAndSearchDropdown';
 import { FolderMenu } from './FolderMenu';
 
 interface NavigationBarProps {
