+++ packages/components/src/internal/components/search/SearchBox.spec.tsx
